+++ SiGeNA/View/src/main/java/sigena/controller/HomeController.java
@@ -1,25 +1,37 @@
-
 package sigena.controller;
 
+import java.io.IOException;
+import java.io.PrintWriter;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
+import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
-import java.io.IOException;
 import java.sql.SQLException;
+import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
 import java.util.List;
 import sigena.model.common.exception.DataInvalidaException;
 import sigena.model.common.exception.DatabaseException;
-import sigena.model.domain.util.Cargo;
+import sigena.model.common.exception.PersistenciaException;
+import sigena.model.domain.Evento;
+import sigena.model.domain.Notificacao;
 import sigena.model.domain.Tarefa;
+import sigena.model.domain.Usuario;
+import sigena.model.domain.util.Cargo;
 import sigena.model.service.FuncionarioService;
+import sigena.model.service.GestaoEventoService;
+import sigena.model.service.GestaoNotificacaoService;
 import sigena.model.service.GestaoTarefaService;
 
-@WebServlet(name = "TarefaController", urlPatterns = {"/TarefaController"})
-public class TarefaController extends Controller {
+@WebServlet(name = "HomeController", urlPatterns = {"/HomeController"})
+public class HomeController extends Controller {
+
+    private GestaoNotificacaoService serviceN = new GestaoNotificacaoService();
+    private GestaoEventoService serviceE = new GestaoEventoService();
 
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
@@ -33,6 +45,24 @@ public class TarefaController extends Controller {
 
         String acao = request.getParameter("acao");
         GestaoTarefaService service = new GestaoTarefaService();
+        Usuario usuario = (Usuario) request.getSession().getAttribute("UsuarioLogado");
+
+        LocalDate hoje = LocalDate.now();
+        LocalDate ultimaExecucao = (LocalDate) request.getSession()
+                .getAttribute("notificacaoExecutada");
+
+        if (ultimaExecucao == null || !ultimaExecucao.equals(hoje)) {
+            try {
+                notificarEventos(usuario);
+                request.getSession().setAttribute("notificacaoExecutada", hoje);
+            } catch (PersistenciaException e) {
+                e.printStackTrace();
+            }
+        }
+
+        List<Notificacao> notificacoes = serviceN.listarPorUsuario(usuario.getId());
+
+        request.setAttribute("notificacoes", notificacoes);
 
         try {
             if ("cadastrar".equals(acao)) {
@@ -97,8 +127,10 @@ public class TarefaController extends Controller {
                     excluir(request, response);
                 case "concluir" ->
                     concluir(request, response);
+                case "lerNotificacao" ->
+                    marcarComoLida(request, response);
                 default ->
-                    response.sendRedirect("TarefaController");
+                    response.sendRedirect("HomeController");
             }
         } catch (Exception e) {
             throw new ServletException(e);
@@ -137,7 +169,7 @@ public class TarefaController extends Controller {
             return;
         }
 
-        response.sendRedirect("TarefaController");
+        response.sendRedirect("HomeController");
     }
 
     private void editar(HttpServletRequest request, HttpServletResponse response)
@@ -159,14 +191,14 @@ public class TarefaController extends Controller {
         try {
             service.editar(id, nome, texto, idDestinatario, dataPConclusao, null, cpfLogado);
         } catch (DataInvalidaException e) {
-            response.sendRedirect("TarefaController?acao=editar&id=" + id);
+            response.sendRedirect("HomeController?acao=editar&id=" + id);
             return;
         } catch (SecurityException e) {
             response.sendError(HttpServletResponse.SC_FORBIDDEN);
             return;
         }
 
-        response.sendRedirect("TarefaController");
+        response.sendRedirect("HomeController");
     }
 
     private void excluir(HttpServletRequest request, HttpServletResponse response)
@@ -190,7 +222,7 @@ public class TarefaController extends Controller {
             return;
         }
 
-        response.sendRedirect("TarefaController");
+        response.sendRedirect("HomeController");
     }
 
     private void concluir(HttpServletRequest request, HttpServletResponse response)
@@ -207,6 +239,27 @@ public class TarefaController extends Controller {
             service.editarConcluida(id, true, cpf);
         }
 
-        response.sendRedirect("TarefaController");
+        response.sendRedirect("HomeController");
     }
+
+    public void notificarEventos(Usuario usuario) throws PersistenciaException, ServletException, IOException {
+        LocalDateTime amanha = LocalDate.now().plusDays(1).atStartOfDay();
+
+        List<Evento> eventos = serviceE.listarEventos(amanha, amanha.with(LocalTime.MAX));
+        
+        for (Evento e : eventos) {
+            if (!serviceN.eventoJaNotificado(usuario.getId(), "Amanhã você tem o evento " + e.getTitulo() + "", amanha)) {
+                Notificacao n = new Notificacao(usuario.getId(), "Amanhã você tem o evento " + e.getTitulo() + "");
+                serviceN.salvar(n);
+            }
+        }
+    }
+
+    private void marcarComoLida(HttpServletRequest request, HttpServletResponse response) throws IOException {
+        Long id = Long.parseLong(request.getParameter("id"));
+        Notificacao n = serviceN.buscarPorId(id);
+        serviceN.marcarComoLida(n);
+        response.sendRedirect("HomeController");
+    }
+
 }
