+++ SiGeNA/View/src/main/webapp/editar-produto.jsp
@@ -27,7 +27,7 @@
 
     <body>
         <header class="topbar">
-            <a href="TarefaController" class="titulo">
+            <a href="HomeController" class="titulo">
                 <img src="IMG's/logoSiGeNA-COR2.png" alt="Logo" class="brand-logo">
                 <span>SiGeNA</span>
             </a>
