+++ SiGeNA/View/src/main/webapp/eventos.jsp
@@ -26,7 +26,7 @@
 </head>
 <body>
 <header class="topbar">
-            <a href="TarefaController" class="titulo">
+            <a href="HomeController" class="titulo">
                 <img src="IMG's/logoSiGeNA-COR2.png" alt="Logo" class="brand-logo">
                 <span>SiGeNA</span>
             </a>
